@@ -14,8 +14,6 @@
 
 from __future__ import annotations
 
-from typing import List
-
 import strawberry
 
 try:
@@ -68,7 +66,7 @@ async def resolve_search(contains: str):
 class Query:
     library: Library = field(resolver=resolve_library)
     hello: str = field(resolver=resolve_hello)
-    search: List[Item] = field(resolver=resolve_search)
+    search: list[Item] = field(resolver=resolve_search)
     echo: str = field(resolver=resolve_echo)
     storage: Storage = field(resolver=resolve_storage)
     error: str | None = field(resolver=resolve_error)
