@@ -14,7 +14,7 @@
 
 from __future__ import annotations
 
-from typing import List, Union
+from typing import Union
 
 import strawberry
 
@@ -56,12 +56,12 @@ class Magazine:
 class Library:
     id: int
     branch: str
-    magazine: List[Magazine]
-    book: List[Book]
+    magazine: list[Magazine]
+    book: list[Book]
 
 
 Item = Union[Book, Magazine]
-Storage = List[str]
+Storage = list[str]
 
 
 authors = [
@@ -138,7 +138,7 @@ def resolve_search(contains: str):
 class Query:
     library: Library = field(resolver=resolve_library)
     hello: str = field(resolver=resolve_hello)
-    search: List[Item] = field(resolver=resolve_search)
+    search: list[Item] = field(resolver=resolve_search)
     echo: str = field(resolver=resolve_echo)
     storage: Storage = field(resolver=resolve_storage)
     error: str | None = field(resolver=resolve_error)
